+++ internal/domain/BTMCustomerNote.go
@@ -11,7 +11,7 @@ type BTMCustomerNote struct {
 	CreatedAt         time.Time        `json:"created_at"`
 	CustomerId        uuid.UUID        `json:"customer_id"`
 	Note              string           `json:"note"`
-	OperationUserId   uint             `json:"operation_user_id"`
+	OperationUserId   int64            `json:"operation_user_id"`
 	OperationUserName string           `json:"operation_user_name"`
 	NoteType          CustomerNoteType `json:"note_type"`
 }
